backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
torch
torchvision
pillow
numpy
google-generativeai
reportlab
matplotlib
opencv-python-headless
